auth-server: server: Move helpers to more discoverable locations
