+++ auth/auth-server/src/main.rs
@@ -18,7 +18,6 @@
 mod bundle_store;
 mod chain_events;
 pub(crate) mod error;
-mod helpers;
 pub mod http_utils;
 mod server;
 mod telemetry;
