+++ auth/auth-server/src/server/api_handlers/key_management.rs
@@ -1,8 +1,8 @@
 //! Handles key management requests
 
-use crate::server::{
-    db::models::NewApiKey,
-    helpers::{aes_encrypt, empty_json_reply},
+use crate::{
+    http_utils::empty_json_reply,
+    server::{db::models::NewApiKey, helpers::aes_encrypt},
 };
 use auth_server_api::CreateApiKeyRequest;
 use bytes::Bytes;
