+++ auth/auth-server/src/server/api_handlers/mod.rs
@@ -12,7 +12,7 @@ use auth_server_api::{
     SponsoredMatchResponse, SponsoredQuoteResponse,
 };
 use bytes::Bytes;
-use http::{HeaderMap, Method, StatusCode};
+use http::{HeaderMap, Method, Response, StatusCode};
 use renegade_api::http::external_match::{
     AssembleExternalMatchRequest, ExternalMatchRequest, ExternalMatchResponse, ExternalOrder,
     ExternalQuoteRequest, ExternalQuoteResponse, MalleableExternalMatchResponse,
@@ -29,11 +29,10 @@ use super::gas_sponsorship::refund_calculation::{
     apply_gas_sponsorship_to_exact_output_amount, remove_gas_sponsorship_from_quote,
     requires_exact_output_amount_update,
 };
-use super::helpers::{
-    generate_quote_uuid, get_sdk_version, log_unsuccessful_relayer_request, overwrite_response_body,
-};
+use super::helpers::generate_quote_uuid;
 use super::Server;
 use crate::error::AuthServerError;
+use crate::http_utils::overwrite_response_body;
 use crate::telemetry::helpers::{
     calculate_implied_price, record_quote_not_found, record_relayer_request_500,
 };
@@ -48,9 +47,47 @@ use crate::telemetry::{
     },
 };
 
+/// The header name for the SDK version
+const SDK_VERSION_HEADER: &str = "x-renegade-sdk-version";
+/// The default SDK version to use if the header is not set
+const SDK_VERSION_DEFAULT: &str = "pre-v0.1.0";
+
+/// Parse the SDK version from the given headers.
+/// If unset or malformed, returns an empty string.
+pub fn get_sdk_version(headers: &HeaderMap) -> String {
+    headers
+        .get(SDK_VERSION_HEADER)
+        .map(|v| v.to_str().unwrap_or_default())
+        .unwrap_or(SDK_VERSION_DEFAULT)
+        .to_string()
+}
+
+/// Log a non-200 response from the relayer for the given request
+pub fn log_unsuccessful_relayer_request(
+    resp: &Response<Bytes>,
+    key_description: &str,
+    path: &str,
+    req_body: &[u8],
+    headers: &HeaderMap,
+) {
+    let status = resp.status();
+    let text = String::from_utf8_lossy(resp.body()).to_string();
+    let req_body = String::from_utf8_lossy(req_body).to_string();
+    let sdk_version = get_sdk_version(headers);
+    warn!(
+        key_description = key_description,
+        path = path,
+        request_body = req_body,
+        sdk_version = sdk_version,
+        "Non-200 response from relayer: {status}: {text}",
+    );
+}
+
 // ---------------
 // | Server Impl |
 // ---------------
+
+// TODO(@joeykraut): Move individual handlers to their own files
 
 /// Handle a proxied request
 impl Server {
