+++ auth/auth-server/src/server/api_handlers/order_book.rs
@@ -4,11 +4,8 @@ use http::{HeaderMap, Method, StatusCode};
 use tracing::instrument;
 use warp::{reject::Rejection, reply::Reply};
 
-use super::Server;
-use crate::{
-    server::helpers::log_unsuccessful_relayer_request,
-    telemetry::helpers::record_relayer_request_500,
-};
+use super::{log_unsuccessful_relayer_request, Server};
+use crate::telemetry::helpers::record_relayer_request_500;
 
 impl Server {
     /// Proxy GET /v0/order_book/depth/:mint to the relayer
